Category module

+++ src/document/document.service.ts
@@ -174,7 +174,7 @@ export class DocumentService {
     voteTime.setDate(voteTime.getDate() + 30);
     docTime.setDate(docTime.getDate() + 37);
 
-    if (body.categoryId === 5 && body.goal) {
+    if (body.categoryId === this.goodsCategoryId && body.goal) {
       docOption = await this.DocumentOptionRepo.save({
         goal: body.goal,
         voteExpire: voteTime,
@@ -214,7 +214,7 @@ export class DocumentService {
     return document;
   }
 
-  async deleteDocument(documentId: number) {
+  async deleteDocument(documentId: number, user) {
     const document = await this.DocumentRepo.findOne({
       where: { id: documentId },
       relations: {
@@ -227,6 +227,9 @@ export class DocumentService {
     });
     if (!document) {
       throw new NotFoundException(`Document with ID ${documentId} not found`);
+    }
+    if (document.author.id !== user.userId) {
+      throw new UnauthorizedException(`Only writer can edit document`);
     }
 
     await this.imageService.deleteDir(`${documentId}`);
@@ -244,7 +247,7 @@ export class DocumentService {
     await this.DocumentRepo.remove(document);
 
     // delete docOption if category is "goods or 5"
-    if (document.category.id === 5) {
+    if (document.category.id === this.goodsCategoryId) {
       const customOption = document.option;
       await this.DocumentOptionRepo.remove(customOption);
     }
@@ -325,8 +328,10 @@ export class DocumentService {
                 );
                 if (directory) {
                   document.images[i].directory = directory;
-                  document.images[i].filename = `${updateDocumentDTO.imageName[i]}`,
-                  await this.ImageRepo.save(document.images[i]);
+                  (document.images[
+                    i
+                  ].filename = `${updateDocumentDTO.imageName[i]}`),
+                    await this.ImageRepo.save(document.images[i]);
                 }
               } else {
                 console.log('image not deleted');
